Rename gsound_context_change_attrs() to set_attributes()
This method just wraps "ca_context_change_props()", which is how
it got its name. But I think the abbreviation of "attribute" is
just confusing, and this is conceptually a "set" operation.

Mind you, calling it "set" like this might lead people to wonder
why there is no corresponding "get"...?